Moved PointyLabelNode to a new file.
